jenkins_one_at_a_time_hash return uint32_t.

+++ libdariadb/storage/bloom_filter.h
@@ -5,10 +5,10 @@
 namespace dariadb {
 namespace storage {
 
-template <class T> uint64_t jenkins_one_at_a_time_hash(T &value) {
+template <class T> uint32_t jenkins_one_at_a_time_hash(T &value) {
   auto key = reinterpret_cast<const char*>(&value);
   auto len = sizeof(value);
-  uint64_t hash, i;
+  uint32_t hash, i;
   for (hash = i = 0; i < len; ++i) {
     hash += key[i];
     hash += (hash << 10);
